Removed a kind specification.

+++ src/opsinputs/opsinputs_cxwriter_mod.F90
@@ -162,7 +162,7 @@ logical                                    :: opsinputs_cxwriter_create
 
 ! Local declarations:
 character(len=:), allocatable              :: string
-integer(kind=c_int)                        :: int
+integer                                    :: int
 logical                                    :: bool
 real(kind=c_double)                        :: double
 logical                                    :: found
